requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
Werkzeug==2.0.1
pymysql==1.0.2
cryptography==3.4.7
python-dotenv==0.19.0